+++ src/samplers/syscall/linux/counts/mod.bpf.c
@@ -39,7 +39,7 @@ struct {
 	__uint(type, BPF_MAP_TYPE_ARRAY);
 	__uint(map_flags, BPF_F_MMAPABLE);
 	__type(key, u32);
-	__type(value, u64);
+	__type(value, u32);
 	__uint(max_entries, MAX_SYSCALL_ID);
 } syscall_lut SEC(".maps");
